Invert check for user approval to avoid nesting
Improve the readablitity of the code by inverting the check for user
approval. This allows us to drop a layer of nesting for a large part
of the script. This is also more consistent with how these sorts of
checks are handled elsewhere in the script.

Also move the cursor down after prompting for the users approval so
that we don't start writing log messages on the same line.